PeeringDB 2020 Satisfaction Survey
This commit contains the text and image for the PeeringDB 2020 Satisfaction Survey blog post. The text is broadly the same as that used by partners who are helping us publicize the survey with a minor modification reflecting that we are not a partner but are the organization running the survey. If approved, this blog post should be made live on 2 November 2020.