Facilities HOWTO, tools page, blog, and index.md reorg
This change adds a facilities HOWTO and moves our tools listing from index.md to a separate page. The listing has been updated with new entries, and a project that is closing removed. There is an accompanying blog post. Also, the full list of presentations has been moved to its own page as it was so long.